Split Utility class to several classes
