+++ app/src/main/java/com/example/android/sunshine/app/utils/DateUtility.java
@@ -1,0 +1,117 @@
+package com.example.android.sunshine.app.utils;
+
+import android.annotation.SuppressLint;
+import android.content.Context;
+import android.text.format.Time;
+
+import com.example.android.sunshine.app.R;
+
+import java.text.SimpleDateFormat;
+
+@SuppressWarnings("deprecation")
+@SuppressLint("SimpleDateFormat")
+public class DateUtility {
+
+    private DateUtility() throws IllegalAccessException {
+        throw new IllegalAccessException("Class shouldn't be instantiated!");
+    }
+
+    /**
+     * Helper method to convert the database representation of the date into something to display
+     * to users.  As classy and polished a user experience as "20140102" is, we can do better.
+     *
+     * @param context      Context to use for resource localization
+     * @param dateInMillis The date in milliseconds
+     * @return a user-friendly representation of the date.
+     */
+    public static String getFriendlyDayString(Context context, long dateInMillis) {
+        // The day string for forecast uses the following logic:
+        // For today: "Today, June 8"
+        // For tomorrow:  "Tomorrow"
+        // For the next 5 days: "Wednesday" (just the day name)
+        // For all days after that: "Mon Jun 8"
+
+        Time time = new Time();
+        time.setToNow();
+        long currentTime = System.currentTimeMillis();
+        int julianDay = Time.getJulianDay(dateInMillis, time.gmtoff);
+        int currentJulianDay = Time.getJulianDay(currentTime, time.gmtoff);
+
+        // If the date we're building the String for is today's date, the format
+        // is "Today, June 24"
+        if (julianDay == currentJulianDay) {
+            String today = context.getString(R.string.today);
+            int formatId = R.string.format_full_friendly_date;
+            return context.getString(
+                    formatId,
+                    today,
+                    getFormattedMonthDay(dateInMillis)
+            );
+        } else if (julianDay < currentJulianDay + 7) {
+            // If the input date is less than a week in the future, just return the day name.
+            return getDayName(context, dateInMillis);
+        } else {
+            // Otherwise, use the form "Mon Jun 3"
+            SimpleDateFormat shortenedDateFormat = new SimpleDateFormat("EEE MMM dd");
+            return shortenedDateFormat.format(dateInMillis);
+        }
+    }
+
+    /**
+     * Converts db date format to the format "Month day", e.g "June 24".
+     *
+     * @param dateInMillis The db formatted date string, expected to be of the form specified
+     *                     in Utility.DATE_FORMAT
+     * @return The day in the form of a string formatted "December 6"
+     */
+    private static String getFormattedMonthDay(long dateInMillis) {
+        Time time = new Time();
+        time.setToNow();
+        SimpleDateFormat monthDayFormat = new SimpleDateFormat("MMMM dd");
+        return monthDayFormat.format(dateInMillis);
+    }
+
+    /**
+     * Helper method to convert the database representation of the date into something to display
+     * to users.  As classy and polished a user experience as "20140102" is, we can do better.
+     *
+     * @param context      Context to use for resource localization
+     * @param dateInMillis The date in milliseconds
+     * @return a user-friendly representation of the date.
+     */
+    public static String getFullFriendlyDayString(Context context, long dateInMillis) {
+
+        String day = getDayName(context, dateInMillis);
+        int formatId = R.string.format_full_friendly_date;
+        return context.getString(
+                formatId,
+                day,
+                getFormattedMonthDay(dateInMillis)
+        );
+    }
+
+    /**
+     * Given a day, returns just the name to use for that day.
+     * E.g "today", "tomorrow", "wednesday".
+     */
+    private static String getDayName(Context context, long dateInMillis) {
+        // If the date is today, return the localized version of "Today" instead of the actual
+        // day name.
+
+        Time t = new Time();
+        t.setToNow();
+        int julianDay = Time.getJulianDay(dateInMillis, t.gmtoff);
+        int currentJulianDay = Time.getJulianDay(System.currentTimeMillis(), t.gmtoff);
+        if (julianDay == currentJulianDay) {
+            return context.getString(R.string.today);
+        } else if (julianDay == currentJulianDay + 1) {
+            return context.getString(R.string.tomorrow);
+        } else {
+            Time time = new Time();
+            time.setToNow();
+            // Otherwise, the format is just the day of the week (e.g "Wednesday".
+            SimpleDateFormat dayFormat = new SimpleDateFormat("EEEE");
+            return dayFormat.format(dateInMillis);
+        }
+    }
+}
